+++ src/widget/widget.rs
@@ -303,11 +303,13 @@ pub struct DefaultWidgetCallbacks {
     pub on_mouse_move: Option<Box<dyn FnMut(&mut dyn Widget, Point, &Vec<WidgetContainer>)>>,
     pub on_mouse_button: Option<Box<dyn FnMut(&mut dyn Widget, Button, bool, &Vec<WidgetContainer>)>>,
     pub on_tick: Option<Box<dyn FnMut(&mut dyn Widget, &Vec<WidgetContainer>)>>,
+    pub on_mouse_bounds: Option<Box<dyn FnMut(&mut dyn Widget, bool, &Vec<WidgetContainer>)>>,
     on_click_populated: bool,
     on_toggle_populated: bool,
     on_mouse_move_populated: bool,
     on_mouse_button_populated: bool,
     on_tick_populated: bool,
+    on_mouse_bounds_populated: bool,
 }
 
 impl DefaultWidgetCallbacks {
@@ -318,11 +320,13 @@ impl DefaultWidgetCallbacks {
             on_mouse_move: None,
             on_mouse_button: None,
             on_tick: None,
+            on_mouse_bounds: None,
             on_click_populated: false,
             on_toggle_populated: false,
             on_mouse_move_populated: false,
             on_mouse_button_populated: false,
             on_tick_populated: false,
+            on_mouse_bounds_populated: false,
         }
     }
 
@@ -384,6 +388,18 @@ impl DefaultWidgetCallbacks {
 
     pub fn has_on_tick(&mut self) -> bool {
         self.on_tick_populated
+    }
+
+    pub fn on_mouse_bounds<F>(&mut self, callback: F)
+        where
+            F: FnMut(&mut dyn Widget, bool, &Vec<WidgetContainer>) + 'static,
+    {
+        self.on_mouse_bounds = Some(Box::new(callback));
+        self.on_mouse_bounds_populated = true;
+    }
+
+    pub fn has_on_mouse_bounds(&mut self) -> bool {
+        self.on_mouse_bounds_populated
     }
 }